# docs/api.yaml
openapi: 3.1.0
info:
  title: Customer and Subscription API
  description: API for managing customers and subscriptions.
  version: 1.0.0

paths:
  /customers:
    post:
      tags:
        - Customers
      summary: Create a new customer
      operationId: createCustomer
      requestBody:
        description: Email address of the customer
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: "customer@example.com"
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    $ref: '#/components/schemas/Customer'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
  /subscriptions:
    post:
      tags:
        - Subscriptions
      summary: Create a new subscription
      operationId: createSubscription
      requestBody:
        description: Price ID for the subscription
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                priceId:
                  type: string
                  example: "price_1JHj8eI6eZbJ2T"
      responses:
        '201':
          description: Subscription created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptionId:
                    type: string
                    example: "sub_1234567890"
                  clientSecret:
                    type: string
                    example: "pi_1JHj8eI6eZbJ2T_secret_ABC"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
    put:
      tags:
        - Subscriptions
      summary: Update existing subscription
      operationId: updateSubscription
      requestBody:
        description: Update subscription with new price lookup key
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscriptionId:
                  type: string
                  description: ID of the subscription to update
                  example: "sub_1234567890"
                newPriceLookupKey:
                  type: string
                  description: Lookup key for the new price
                  example: "new_price_key"
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '404':
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
  /prices:
    get:
      tags:
        - Prices
      summary: Retrieve a list of prices
      operationId: listPrices
      responses:
        '200':
          description: List of prices retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  publishableKey:
                    type: string
                    example: "pk_test_1234567890"
                  prices:
                    type: array
                    items:
                      $ref: '#/components/schemas/Price'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ErrorResponse'
components:
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "invalid request"
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: string
          example: "cus_1234567890"
        email:
          type: string
          format: email
          example: "customer@example.com"
        example: 1625080400
    Subscription:
      type: object
      properties:
        id:
          type: string
          example: "sub_1234567890"
        application:
          type: object
          properties:
            id:
              type: string
              example: "ca_1234567890"
        application_fee_percent:
          type: number
          format: float
          example: 2.5
        automatic_tax:
          type: object
          properties:
            enabled:
              type: boolean
              example: false
        billing_cycle_anchor:
          type: integer
          example: 1627682400
        billing_thresholds:
          type: object
          properties:
            amount_gte:
              type: integer
              example: 1000
            reset_billing_cycle_anchor:
              type: boolean
              example: false
        cancel_at:
          type: integer
          example: 1627682400
        cancel_at_period_end:
          type: boolean
          example: false
        canceled_at:
          type: integer
          example: 1627682400
        collection_method:
          type: string
          enum: [charge_automatically, send_invoice]
          example: charge_automatically
        created:
          type: integer
          example: 1627682400
        currency:
          type: string
          example: "usd"
        current_period_end:
          type: integer
          example: 1627682400
        current_period_start:
          type: integer
          example: 1625080400
        customer:
          type: object
          properties:
            id:
              type: string
              example: "cus_1234567890"
        days_until_due:
          type: integer
          example: 30
        default_payment_method:
          type: object
          properties:
            id:
              type: string
              example: "pm_1234567890"
        default_source:
          type: object
          properties:
            id:
              type: string
              example: "card_1234567890"
        default_tax_rates:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "txr_1234567890"
        description:
          type: string
          example: "Monthly subscription for premium plan"
        discount:
          type: object
          properties:
            id:
              type: string
              example: "di_1234567890"
        ended_at:
          type: integer
          example: 1627682400
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: "si_1234567890"
              price:
                type: object
                properties:
                  id:
                    type: string
                    example: "price_1JHj8eI6eZbJ2T"
        latest_invoice:
          type: object
          properties:
            id:
              type: string
              example: "in_1234567890"
        livemode:
          type: boolean
          example: false
        metadata:
          type: object
          additionalProperties:
            type: string
        next_pending_invoice_item_invoice:
          type: integer
          example: 1627682400
        object:
          type: string
          example: "subscription"
        on_behalf_of:
          type: object
          properties:
            id:
              type: string
              example: "acct_1234567890"
        pause_collection:
          type: object
          properties:
            behavior:
              type: string
              enum: [keep_as_draft, mark_uncollectible, void]
              example: keep_as_draft
        payment_settings:
          type: object
          properties:
            payment_method_options:
              type: object
              additionalProperties:
                type: object
            payment_method_types:
              type: array
              items:
                type: string
                example: "card"
        pending_invoice_item_interval:
          type: object
          properties:
            interval:
              type: string
              enum: [day, week, month, year]
              example: month
            interval_count:
              type: integer
              example: 1
        pending_setup_intent:
          type: object
          properties:
            id:
              type: string
              example: "seti_1234567890"
        pending_update:
          type: object
          properties:
            id:
              type: string
              example: "sub_upd_1234567890"
        plan:
          type: object
          properties:
            id:
              type: string
              example: "plan_1234567890"
        quantity:
          type: integer
          example: 1
        schedule:
          type: object
          properties:
            id:
              type: string
              example: "sub_sched_1234567890"
        start_date:
          type: integer
          example: 1625080400
        status:
          type: string
          enum: [incomplete, incomplete_expired, trialing, active, past_due, canceled, unpaid]
          example: canceled
        test_clock:
          type: object
          properties:
            id:
              type: string
              example: "clock_1234567890"
        transfer_data:
          type: object
          properties:
            destination:
              type: string
              example: "acct_1234567890"
        trial_end:
          type: integer
          example: 1627682400
        trial_start:
          type: integer
          example: 1625080400
    Price:
      type: object
      properties:
        id:
          type: string
          example: "price_1JHj8eI6eZbJ2T"
        object:
          type: string
          example: "price"
        active:
          type: boolean
          example: true
        billing_scheme:
          type: string
          example: "per_unit"
        created:
          type: integer
          example: 1627682400
        currency:
          type: string
          example: "usd"
        lookup_key:
          type: string
          example: "sample_basic"
        nickname:
          type: string
          example: "Basic Plan"
        product:
          type: string
          example: "prod_1234567890"
        recurring:
          type: object
          properties:
            interval:
              type: string
              example: "month"
            interval_count:
              type: integer
              example: 1
        unit_amount:
          type: integer
          example: 2000
        unit_amount_decimal:
          type: string
          example: "2000"